requirements.txt
flask
python-dotenv
openai
pinecone-client
twilio
requests
